requirements.txt
# Requirements detected from project imports
# Note: For CUDA-enabled installs, follow official PyTorch instructions at https://pytorch.org/get-started/locally/

torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=8.0.0
opencv-python>=4.5.3.56
matplotlib>=3.4.0
scikit-learn>=1.0.0
tqdm>=4.64.0
notebook>=6.0.0
